Extract corpus generation from engine
At the moment we build the corpus from scratch every time we create a
new GensimEngine instance.

Move this into a separate class so that we can load a corpus we've saved
already.

The `from_documents` factory method is equivalent to the old way of
initialising the engine.

The `from_experiment` factory method uses a saved corpus and dictionary,
using the experiment name we generated based on the date.
